Refactor modules out of core wrapper
