+++ src/assertions/OrderAssertions.ts
@@ -27,7 +27,7 @@ import { CommonAssertions } from './CommonAssertions';
 import { SchemaAssertions } from './SchemaAssertions';
 import { ERC20Assertions } from './ERC20Assertions';
 import { SetTokenAssertions } from './SetTokenAssertions';
-import { CoreWrapper } from '../wrappers';
+import { CoreWrapper, IssuanceOrderModuleWrapper } from '../wrappers';
 import { NULL_ADDRESS, ZERO } from '../constants';
 import { BigNumber, calculatePartialAmount } from '../util';
 import {
@@ -42,15 +42,17 @@ import {
 export class OrderAssertions {
   private web3: Web3;
   private core: CoreWrapper;
+  private issuanceOrderModuleWrapper: IssuanceOrderModuleWrapper;
   private erc20Assertions: ERC20Assertions;
   private schemaAssertions: SchemaAssertions;
   private coreAssertions: CoreAssertions;
   private commonAssertions: CommonAssertions;
   private setTokenAssertions: SetTokenAssertions;
 
-  constructor(web3: Web3, coreWrapper: CoreWrapper) {
+  constructor(web3: Web3, coreWrapper: CoreWrapper, issuanceOrderModuleWrapper: IssuanceOrderModuleWrapper) {
     this.web3 = web3;
     this.core = coreWrapper;
+    this.issuanceOrderModuleWrapper = issuanceOrderModuleWrapper;
     this.erc20Assertions = new ERC20Assertions(web3);
     this.schemaAssertions = new SchemaAssertions();
     this.coreAssertions = new CoreAssertions(web3);
@@ -441,8 +443,8 @@ export class OrderAssertions {
   private async calculateFillableQuantity(signedIssuanceOrder: SignedIssuanceOrder): Promise<BigNumber> {
     const issuanceOrder: IssuanceOrder = _.omit(signedIssuanceOrder, 'signature');
     const orderHash = SetProtocolUtils.hashOrderHex(issuanceOrder);
-    const filledAmount = await this.core.orderFills(orderHash);
-    const cancelledAmount = await this.core.orderCancels(orderHash);
+    const filledAmount = await this.issuanceOrderModuleWrapper.orderFills(orderHash);
+    const cancelledAmount = await this.issuanceOrderModuleWrapper.orderCancels(orderHash);
 
     return issuanceOrder.quantity.sub(filledAmount).sub(cancelledAmount);
   }
