+++ test/integration/api/IssuanceAPI.spec.ts
@@ -114,9 +114,8 @@ describe('IssuanceAPI', () => {
       transferProxy.address,
       vault.address,
       rebalanceAuctionModule.address,
-      issuanceOrderModule.address
     );
-    const assertions = new Assertions(web3, coreWrapper);
+    const assertions = new Assertions(web3);
     issuanceAPI = new IssuanceAPI(web3, coreWrapper, assertions);
 
     componentTokens = await deployTokensAsync(3, web3);
