+++ test/integration/api/OrderAPI.spec.ts
@@ -45,7 +45,7 @@ import {
 
 import { BigNumber, SignatureUtils } from '@src/util';
 import ChaiSetup from '@test/helpers/chaiSetup';
-import { CoreWrapper } from '@src/wrappers';
+import { CoreWrapper, IssuanceOrderModuleWrapper } from '@src/wrappers';
 import { DEFAULT_ACCOUNT, ACCOUNTS } from '@src/constants/accounts';
 import { OrderAPI } from '@src/api';
 import {
@@ -96,6 +96,7 @@ describe('OrderAPI', () => {
   let rebalanceAuctionModule: RebalanceAuctionModuleContract;
 
   let coreWrapper: CoreWrapper;
+  let issuanceOrderModuleWrapper: IssuanceOrderModuleWrapper;
   let ordersAPI: OrderAPI;
 
   beforeEach(async () => {
@@ -124,10 +125,22 @@ describe('OrderAPI', () => {
       transferProxy.address,
       vault.address,
       rebalanceAuctionModule.address,
+    );
+
+    issuanceOrderModuleWrapper = new IssuanceOrderModuleWrapper(
+      web3,
       issuanceOrderModule.address
     );
-    const assertions = new Assertions(web3, coreWrapper);
-    ordersAPI = new OrderAPI(web3, coreWrapper, assertions, kyberNetworkWrapper.address);
+
+    const assertions = new Assertions(web3);
+    assertions.setOrderAssertions(web3, coreWrapper, issuanceOrderModuleWrapper);
+    ordersAPI = new OrderAPI(
+      web3,
+      assertions,
+      issuanceOrderModule.address,
+      kyberNetworkWrapper.address,
+      vault.address,
+    );
   });
 
   afterEach(async () => {
