+++ test/integration/api/SetTokenAPI.spec.ts
@@ -91,10 +91,9 @@ describe('SetTokenAPI', () => {
       transferProxy.address,
       vault.address,
       rebalanceAuctionModule.address,
-      issuanceOrderModule.address
     );
 
-    const assertions = new Assertions(web3, coreWrapper);
+    const assertions = new Assertions(web3);
     setTokenAPI = new SetTokenAPI(web3, assertions);
   });
